#11: Restructure files
- Split into api and backend
