+++ api/user_type.go
@@ -1,4 +1,4 @@
-package main
+package api
 
 import (
 	"github.com/graphql-go/graphql"
