+++ backend/user.go
@@ -1,4 +1,4 @@
-package main
+package backend
 
 type User struct {
 	ID    int    `json:"id"`
